feat: Moves the jibri session attach away from mod_muc_events.
Avoids attaching the hook to all stanzas to all hosts and components.
As this was part of mod_muc_events, it is enabled/disabled together with it.

We also not loaded under internal component which is not necessary and avoids multiple checks in events for that prefix.
